fix: 옵션 리포지토리 id 속성 불일치 해결

+++ src/main/java/gift/product/repository/OptionRepository.java
@@ -5,6 +5,6 @@ import gift.product.entity.Option;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface OptionRepository extends JpaRepository<Option, Integer> {
+public interface OptionRepository extends JpaRepository<Option, Long> {
 
 }
